Thought Process
Turning event from Turn JS handles the keydown events as well as the
the click event that triggers turning event. Both handled by helper
functions. Thought process was mostly in terms of how slowly it load if
we just had a ton of divs waiting from the start. Instead we create two
divs each page turn. Three are being considered, two of which will load
and the third waiting for next page turn to load.